fix(project): 修复bug
修复选择照片的bug，添加了比赛模块和打卡模块

+++ module_community/src/main/java/com/example/module_community/ui/CreatePostFragment.java
@@ -5,12 +5,12 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.lifecycle.ViewModelProvider;
-import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -42,6 +42,7 @@ import java.util.List;
  * @Version 1.0
  */
 public class CreatePostFragment extends BaseFragment<FragmentCreatePostBinding> {
+    private CommunityViewModel mViewModel;
 
     public static CreatePostFragment newInstance() {
         return new CreatePostFragment();
@@ -66,19 +67,17 @@ public class CreatePostFragment extends BaseFragment<FragmentCreatePostBinding> 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        CommunityViewModel mViewModel = new ViewModelProvider(this).get(CommunityViewModel.class);
-        NavController navController = NavHostFragment.findNavController(this);
+        mViewModel = new ViewModelProvider(this).get(CommunityViewModel.class);
 
         List<ChoosePicItem> list = new ArrayList<>();
         PostRecyclerviewAdapter adapter = new PostRecyclerviewAdapter();
         binding.recyclerViewPic.setLayoutManager(new GridLayoutManager(getActivity(), 1, LinearLayoutManager.HORIZONTAL, false));
         binding.recyclerViewPic.setAdapter(adapter);
 
-
         binding.recyclerViewPic.addOnItemTouchListener(new MyOnItemTouchListener(getActivity(), binding.recyclerViewPic, new MyOnItemTouchListener.OnItemClickListener() {
             @Override
             public void onItemClick(View view, int position) {
-                if (position == list.size()) {
+                if (position == adapter.getList().size()) {
                     list.clear();
                     PictureSelector.create(getActivity()).openGallery(SelectMimeType.ofImage()).setImageEngine(GlideEngine.createGlideEngine()).forResult(new OnResultCallbackListener<LocalMedia>() {
                         @Override
@@ -107,18 +106,29 @@ public class CreatePostFragment extends BaseFragment<FragmentCreatePostBinding> 
 
             @Override
             public void onLongItemClick(View view, int position) {
-                MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(getActivity());
-                builder.setTitle("提示").setMessage("是否删除？");
-                builder.setNeutralButton("close", (dialog, which) -> {
-                });
-                builder.setPositiveButton("删除", (dialog, which) -> {
-                    adapter.deleteItem(position);
-                });
-                AlertDialog dialog = builder.create();
-                dialog.show();
+                if (position != adapter.getList().size()) {
+                    binding.recyclerViewPic.smoothScrollToPosition(adapter.getList().size());
+                    MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(getActivity());
+                    builder.setTitle("提示").setMessage("是否删除？");
+                    builder.setNeutralButton("close", (dialog, which) -> {
+                    });
+                    builder.setPositiveButton("删除", (dialog, which) -> {
+                        adapter.deleteItem(position);
+                    });
+                    AlertDialog dialog = builder.create();
+                    dialog.show();
+                }
             }
         }));
-
+        mViewModel.getPostResultMutableLiveData().observe(getViewLifecycleOwner(), result -> {
+            if (result == null) {
+                return;
+            }
+            if (result.isSuccess()) {
+                Toast.makeText(getActivity(), "发布成功", Toast.LENGTH_SHORT).show();
+                NavHostFragment.findNavController(this).popBackStack();
+            }
+        });
         binding.chipPost.setOnClickListener(v -> {
             mViewModel.postData(MyMMkv.getMyDefaultMMkv().getString("token", null), item2File(adapter.getList()), binding.editTextContent.getText().toString());
         });
